# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger",
        "version": "0.0.1"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "operationId": "dd51a484b7fa0f97c7680fdbde14cf5f",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "username": {
                                                "type": "string"
                                            },
                                            "password": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "username": "industria",
                                    "password": "industria"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Autenticado"
                    },
                    "422": {
                        "description": "UNPROCESSABLE CONTENT"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "operationId": "df9bdc6c7a6d9f20b983526b2ec0f698",
                "responses": {
                    "200": {
                        "description": "Cierre de session"
                    },
                    "422": {
                        "description": "UNPROCESSABLE CONTENT"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/auth/refresh": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "operationId": "30c20c52f5f72828d8c531fd7c9ea3e7",
                "responses": {
                    "200": {
                        "description": "Autenticado"
                    },
                    "422": {
                        "description": "UNPROCESSABLE CONTENT"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/auth/reconection": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "Reconection",
                "operationId": "bda651307266595f335692043f7f8b57",
                "requestBody": {
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reconection"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/me": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Mi perfil",
                "operationId": "8367feff40ba10e43187792cad3132c1",
                "requestBody": {
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "Perfil"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/industry/me": {
            "get": {
                "tags": [
                    "industry"
                ],
                "summary": "Mi perfil",
                "operationId": "e47d27dbc02c29a2b3fd0ddc1d794e0b",
                "requestBody": {
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "Perfil"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Ingrese el token en formato \"Bearer cadenatoken\"",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}